+++ pymia/evaluation/metric/__init__.py
@@ -1,5 +1,5 @@
-from .base import (ConfusionMatrix, IMetric, IConfusionMatrixMetric, ISimpleITKImageMetric, INumpyArrayMetric,
-                   Information)
+from .base import (ConfusionMatrix, Distances, IMetric, IConfusionMatrixMetric, IDistanceMetric, ISimpleITKImageMetric,
+                   INumpyArrayMetric, Information)
 from .metric import (get_all_segmentation_metrics, get_all_regression_metrics, get_overlap_metrics, get_distance_metrics,
                      get_classical_metrics)
 
@@ -10,5 +10,5 @@ from .segmentation import (Accuracy, AdjustedRandIndex, AreaUnderCurve, AverageD
                            GlobalConsistencyError, GroundTruthArea, GroundTruthVolume, HausdorffDistance,
                            InterclassCorrelation, JaccardCoefficient, MahalanobisDistance, MutualInformation, Precision,
                            ProbabilisticDistance, RandIndex, Recall, SegmentationArea, SegmentationVolume,
-                           Sensitivity, Specificity, TrueNegative, TruePositive,
+                           Sensitivity, Specificity, SurfaceOverlap, SurfaceDiceOverlap, TrueNegative, TruePositive,
                            VariationOfInformation, VolumeSimilarity)